+++ webui/src/store/effector/cluster.js
@@ -2,6 +2,10 @@
 import { createEvent } from 'effector'
 import { type ClusterPageState } from 'src/store/reducers/clusterPage.reducer';
 
+export type ClusterDisableServersSuggestion = {
+  uuid: string
+}
+
 export type ClusterRefineURISuggestion = {
   uri_new: string,
   uri_old: string,
@@ -9,7 +13,8 @@ export type ClusterRefineURISuggestion = {
 };
 
 export type ClusterSuggestions = {
-  refine_uri?: ClusterRefineURISuggestion[]
+  refine_uri?: ClusterRefineURISuggestion[],
+  disable_servers?: ClusterDisableServersSuggestion[]
 };
 
 export type ClusterState = {