+++ webui/src/store/effector/clusterSuggestions.js
@@ -4,25 +4,29 @@ import {
   createEvent,
   createEffect,
   createStoreObject,
-  sample
+  sample,
+  combine
 } from 'effector';
 import type { Effect, Store } from 'effector';
 import { equals } from 'ramda';
-import { editServers } from '../request/clusterPage.requests';
+import { disableServers, editServers } from '../request/clusterPage.requests';
 import { getErrorMessage } from '../../api';
 import {
   clusterPageMount,
   statsResponseSuccess,
   statsResponseError,
-  type ClusterRefineURISuggestion
+  type ClusterRefineURISuggestion,
+  type ClusterDisableServersSuggestion
 } from './cluster';
 import type { EditServerInput } from 'src/generated/graphql-typing';
 
-export const applyClick = createEvent<mixed>('apply click');
-export const detailsClick = createEvent<mixed>('details modal click');
+export const advertiseURIApplyClick = createEvent<mixed>('advertise URI apply click');
+export const advertiseURIDetailsClick = createEvent<mixed>('advertise URI details click');
+export const disableServersApplyClick = createEvent<mixed>('disable servers apply click');
+export const disableServersDetailsClick = createEvent<mixed>('disable servers details click');
 export const detailsClose = createEvent<mixed>('details modal close');
 
-export const submitChangesFx: Effect<
+const submitAdvertiseURIFx: Effect<
   Array<EditServerInput>,
   void,
   Error
@@ -31,7 +35,16 @@ export const submitChangesFx: Effect<
   { handler: servers => editServers(servers) }
 );
 
-export const $advertiseURISuggestions: Store<?ClusterRefineURISuggestion[]> = createStore(null)
+const submitDisableServersFx: Effect<
+  Array<string>,
+  void,
+  Error
+> = createEffect(
+  'submit disable servers',
+  { handler: uuids => disableServers(uuids) }
+);
+
+export const $advertiseURISuggestion: Store<?ClusterRefineURISuggestion[]> = createStore(null)
   .on(
     statsResponseSuccess,
     (prev, { suggestions }) => {
@@ -42,33 +55,78 @@ export const $advertiseURISuggestions: Store<?ClusterRefineURISuggestion[]> = cr
   .reset(statsResponseError)
   .reset(clusterPageMount);
 
-export const $advertisePanelVisible: Store<bool> = $advertiseURISuggestions.map(
-  refine_uri => !!refine_uri
+export const $disableServersSuggestion: Store<?ClusterDisableServersSuggestion[]> = createStore(null)
+  .on(
+    statsResponseSuccess,
+    (prev, { suggestions }) => {
+      const next = (suggestions && suggestions.disable_servers) || null;
+      return equals(prev, next) ? prev : next;
+    }
+  )
+  .reset(statsResponseError)
+  .reset(clusterPageMount);
+
+export const $panelsVisibility: Store<{ advertiseURI: bool, disableServers: bool }> = combine(
+  {
+    advertiseURI: $advertiseURISuggestion,
+    disableServers: $disableServersSuggestion
+  },
+  ({ advertiseURI, disableServers }) => ({
+    advertiseURI: !!advertiseURI,
+    disableServers: !!(disableServers && disableServers.length)
+  })
 );
 
-export const $advertiseModalVisible: Store<bool> = createStore(false)
-  .on(detailsClick, () => true)
-  .reset(submitChangesFx.done)
+export const $advertiseURIModalVisible: Store<bool> = createStore(false)
+  .on(advertiseURIDetailsClick, () => true)
+  .reset(submitAdvertiseURIFx.done)
   .reset(detailsClose)
   .reset(clusterPageMount);
 
-export const $error: Store<?string> = createStore(null)
-  .on(submitChangesFx.failData, (_, error) => getErrorMessage(error))
-  .reset(submitChangesFx)
-  .reset(submitChangesFx.done)
+export const $disableServerModalVisible: Store<bool> = createStore(false)
+  .on(disableServersDetailsClick, () => true)
+  .reset(submitDisableServersFx.done)
   .reset(detailsClose)
   .reset(clusterPageMount);
 
-export const $advertiseModal = createStoreObject({
-  visible: $advertiseModalVisible,
-  suggestions: $advertiseURISuggestions,
-  error: $error,
-  pending: submitChangesFx.pending
+export const $advertiseURIError: Store<?string> = createStore(null)
+  .on(submitAdvertiseURIFx.failData, (_, error) => getErrorMessage(error))
+  .reset(submitAdvertiseURIFx)
+  .reset(submitAdvertiseURIFx.done)
+  .reset(detailsClose)
+  .reset(clusterPageMount);
+
+export const $disableServersError: Store<?string> = createStore(null)
+  .on(submitDisableServersFx.failData, (_, error) => getErrorMessage(error))
+  .reset(submitDisableServersFx)
+  .reset(submitDisableServersFx.done)
+  .reset(detailsClose)
+  .reset(clusterPageMount);
+
+export const $advertiseURIModal = createStoreObject({
+  visible: $advertiseURIModalVisible,
+  suggestions: $advertiseURISuggestion,
+  error: $advertiseURIError,
+  pending: submitAdvertiseURIFx.pending
+})
+
+export const $disableServersModal = createStoreObject({
+  visible: $disableServerModalVisible,
+  suggestions: $disableServersSuggestion,
+  error: $disableServersError,
+  pending: submitDisableServersFx.pending
 })
 
 sample({
-  source: $advertiseURISuggestions,
-  clock: applyClick,
+  source: $advertiseURISuggestion,
+  clock: advertiseURIApplyClick,
   fn: servers => (servers || []).map(({ uuid, uri_new: uri }) => ({ uuid, uri })),
-  target: submitChangesFx
+  target: submitAdvertiseURIFx
+});
+
+sample({
+  source: $disableServersSuggestion,
+  clock: disableServersApplyClick,
+  fn: servers => (servers || []).map(({ uuid }) => uuid),
+  target: submitDisableServersFx
 });